net/mlx5e: Move policy delete to run as work
Commit citied in Fixes line avoided the mlx5 command interface limitation
of not-being able to run in atomic context. Unfortunately, in some of policy
releases flows, there won't call to xfrm_dev_policy_delete() at all as it
is done through xfrm_put_pol().

The xfrm_put_pol() can run both in atomic and non-atomic context, so the fixes
patch should be rewritten as a preparation to XFRM core fix.

Fixes: 4a5b8981d3cc ("net/mlx5e: Don't delay release of hardware objects")
Signed-off-by: Leon Romanovsky <leonro@nvidia.com>