net/mlx5e: Close PF netdev when unload driver
When eswitch mode is changed from switchdev to legacy, and vice versa,
PF (or PF rep) netdev is not unregistered. So NETDEV_UNREGISTER event
is not sent to xfrm module to flush the offloaded IPSec states and
policies.

The forbidden of eswitch mode change if there is any IPSec rule
offloaded, is already added before, but users still can do that by
unloading driver, which changes the mode to legacy first, if it's in
switchdev mode.

So close PF rep netdev while unloading driver, NETDEV_DOWN event is
sent to xfrm, and the flushing is also triggered, just like what
NETDEV_UNREGISTER event does.

Fixes: 366e46242b8e ("net/mlx5e: Make IPsec offload work together with eswitch and TC")
Signed-off-by: Jianbo Liu <jianbol@nvidia.com>
Signed-off-by: Leon Romanovsky <leonro@nvidia.com>